requirements.txt
pandas == 0.24.2
numpy >= 1.20.1
scipy
pillow
